build(bazel): Add jest testing macro to barista.
Added a jest-runner for bazel with a custom resolver to the repository.
The custom resolver is needed to resolve the path aliases to the libraries.
Furthermore it has to provide absolute paths as jest cannot scope with
symlinks that are generated by bazel.

Provide a view engine compilation ng_module and one for compiling with ivy.
Jest tests have to use the view engine compile as the TestBed will always use
VE.